chore(macos): keep Rust diagnostics in a log file and clear audit alerts
A Finder launch sent stderr nowhere, so engine diagnostics were lost. Unless
stderr is a terminal, it now goes to ~/Library/Logs/VAYRA/vayra.log, keeping
one previous generation past 10 MiB.

Drop a same-type cast that failed strict clippy, refresh browserslist, js-yaml
and baseline-browser-mapping in the lockfile, and move vitest to 4.1.11 for the
@vitest/mocker path traversal advisory. pnpm audit reports no vulnerability.

+++ src-tauri/src/download.rs
@@ -158,7 +158,7 @@ pub async fn download_available_space(path: String) -> Result<u64, String> {
                 return Err("Storage information unavailable".to_string());
             }
             let info = unsafe { info.assume_init() };
-            Ok((info.f_bavail as u64).saturating_mul(info.f_frsize as u64))
+            Ok((info.f_bavail as u64).saturating_mul(info.f_frsize))
         }).await.map_err(|_| "Storage information unavailable".to_string())?
     }
     #[cfg(not(target_os = "macos"))]
